move parsers.js to pkg/lib, adjust import path in users.js
